Added expense list to Budget view.  Molded and styled all budgets.  Scripts to hide 0 budgets. Scripts to sort expenses by budget.  Add MostRecent expense to add expense

+++ Models/Home/Budget/BudgetV.cs
@@ -10,7 +10,7 @@ namespace TestBudgeting.Models.Home.Budget
         public double BudgetAmount { get; set; }
         public string DistinctBudgets { get; set; }
         public int CurrentMonth { get; set; } = DateTime.Now.Month;
-        public int TotalSpent { get; set; }
+        public double TotalSpent { get; set; }
         public double IdealSpendage { get; set; }
